auth and authorization protection for actions

+++ src/features/ticket/actions/delete-ticket.ts
@@ -4,10 +4,23 @@ import { ticketsPath } from "@/paths";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 import { setCookieByKey } from "@/actions/cookies";
-import { formErrorToActionState } from "@/components/form/util/to-action-state";
+import {
+  formErrorToActionState,
+  toActionState,
+} from "@/components/form/util/to-action-state";
+import { getAuthOrRedirect } from "@/features/auth/queries/get-auth-or-redirect";
+import { isOwner } from "@/features/auth/utils/is-owner";
 
 export const deleteTicket = async (id: string) => {
+  const { user } = await getAuthOrRedirect();
   try {
+    const ticket = await prisma.ticket.findUnique({
+      where: { id },
+    });
+
+    if (!ticket || !isOwner(user, ticket)) {
+      return toActionState("ERROR", "Not Authorized");
+    }
     await prisma.ticket.delete({
       where: {
         id,
